requirements.txt
asyncio~=3.4.3
python-dotenv~=1.0.1
openai~=1.50.2
pypdf2~=3.0.1
langchain~=0.3.1
langchain-core~=0.3.6
langchain-openai~=0.2.1
